Make the first exp of retry with backoff be 0
Context:
--------

Previously the first exponent used to calculate the first waiting
time when retrying using the expoential backoff algorithm with
full jitter was 1, i.e. 2**1 was used instead of 2**0. This made
the first waiting time longer (only potentially because of jitter)
than what it's supposed to be.

Also previously the tests in `retry_test.go` that relied on
`patternTimer` were not actually testing anything, since the `Run`
method of such struct preventing the recording of the waiting times
because `prevExecution` was always equal to time.Time{}.

This PR fixes the issues mentioned above.